components/requirements.txt
openai-whisper
language_tool_python
openai
pyttsx3
gtts
sounddevice
soundfile
python-dotenv
aiohttp
python-socketio
ffmpeg-python